backend/whisper/requirements.txt
whisperx
psycopg2-binary
python-dotenv
beautifulsoup4
ffmpeg-python
uvicorn
fastapi
Pillow
imagehash